refactor(objectives): move Finding synthesis into analysis
Why:
- Cross-paper result construction and its bounded model judgment were split across a root service, generic prompt and schema modules, and the shared response client.

What:
- Co-locate deterministic Finding synthesis and the assertion-strength judgment contract under the analysis stage, then remove the obsolete shared contracts and old import path.

Impact:
- Result-set grouping, Evidence bindings, repair budgets, Finding publication, persistence, and public contracts remain unchanged.

Verification:
- All 169 Finding, structured-response, and confirmed-analysis workflow tests passed, along with all 6 collection-build integration tests.

Refs:
- ISSUE: #330
