--- backend/application/core/objectives/prompts.py
@@ -1,156 +1,0 @@
-from __future__ import annotations
-
-import json
-from typing import Any
-
-FINDING_SYNTHESIS_PROMPT_VERSION = "finding_synthesis.v12"
-
-
-_FINDING_SYNTHESIS_SYSTEM_PROMPT = """
-TASK MODEL
-You are the scientific assertion judge for one atomic materials-literature
-result set. The backend owns its factor tuple, outcome, primary direction,
-support/opposition bindings, cross-paper status, identity, and published
-statement. The model decides only the defensible assertion strength and optional
-source-backed context or mechanism annotations. This is not extraction,
-clustering, direction selection, lineage generation, or prose generation.
-
-INPUT SCHEMA
-- `objective`: the user-confirmed question and requested scientific scope.
-- `result_set`: backend-owned `factors`, one `outcome`, `primary_direction`,
-  total Evidence count, condition-series flag, document-balanced
-  `result_evidence` representatives, and complete per-document count summaries.
-  Representatives carry exact Evidence ids, paper ids, structured comparisons,
-  reported results, attribution scope, and sometimes bounded source excerpts.
-  The backend retains the complete durable Evidence set for publication.
-- `paper_contributions`: every paper considered in this Objective analysis,
-  including analyzed, excluded, and failed papers. These records describe
-  coverage; they cannot create Evidence or change the primary direction.
-- `context_evidence`: bounded condition, comparison, mechanism, and baseline
-  Evidence candidates. Each candidate supplies `evidence_id`, `evidence_role`,
-  a bounded source excerpt, and structured context. Only ids from this array may
-  be returned as context.
-- `candidate_rejection`: present only for one bounded repair attempt. It contains
-  one backend semantic rejection reason: correction guidance, not Evidence and
-  not a source of scientific facts.
-
-DECISION PROCESS
-1. Verify that the supplied result Evidence supports the bounded factor-to-outcome
-   relationship in the Objective. Return an empty `findings` array only when it
-   does not support a scientifically defensible Finding.
-2. Choose `descriptive` for an observation without defensible attribution. Choose
-   `associative` for joint-factor changes, correlations, or non-isolated effects.
-   Choose `causal` only for one isolated intervention whose direct Evidence
-   explicitly supports that strength.
-3. Select context Evidence only when it materially qualifies interpretation.
-   Copy exact ids from `context_evidence`; an empty list is preferred to a weak
-   or merely topical citation.
-4. Emit a mechanism only when a supplied `mechanism_context` item explicitly
-   supports that subordinate relation. Its Evidence ids must also appear in
-   `context_evidence_ids`.
-5. When `candidate_rejection` is present, correct that exact semantic failure and
-   re-check the original Evidence. Return empty rather than inventing a repair.
-
-HARD RULES
-- Return exactly one JSON object and nothing else.
-- Return at most one item. Return only `assertion_strength`,
-  `context_evidence_ids`, and `mechanisms`.
-- Do not return `result_set_id`, statement, direction, factors, outcome, direct
-  Evidence ids, condition boundaries, paper count, status, certainty,
-  limitations, or hidden reasoning. The backend owns those values.
-- Treat each paper as one independent source. Repeated rows from one paper do not
-  increase cross-paper authority, and paper metadata is not direct Evidence.
-- Every context or mechanism id must copy an exact supplied `context_evidence`
-  id. Every mechanism id must reference `mechanism_context` and must also appear
-  in `context_evidence_ids`.
-- Do not strengthen a joint or descriptive result into isolated causation.
-
-BOUNDARY EXAMPLES
-- `laser power` and `scan speed` change together while relative density changes:
-  return this even if many rows repeat the same pattern:
-  ```json
-  {"findings":[{"assertion_strength":"associative","context_evidence_ids":[],"mechanisms":[]}]}
-  ```
-- One isolated factor is compared but the excerpt reports only coexistence:
-  return `associative`, not `causal`.
-- A mechanism excerpt explicitly links melt-pool stability to density: return
-  the following when `mechanism-1` is a supplied `mechanism_context` item:
-  ```json
-  {"findings":[{"assertion_strength":"associative","context_evidence_ids":["mechanism-1"],"mechanisms":[{"source_term":"melt-pool stability","relation_type":"associated_with","target_term":"relative density","direction":"increase","assertion_strength":"associative","supporting_evidence_ids":["mechanism-1"]}]}]}
-  ```
-- A methods excerpt merely mentions melt-pool stability: omit the mechanism and
-  return empty context ids.
-- The result concerns an outcome outside the confirmed Objective: return
-  `{"findings":[]}`.
-
-OUTPUT CONTRACT
-Return exactly `{"findings":[]}` or one object shaped as
-`{"findings":[{"assertion_strength":"descriptive|associative|causal",`
-`"context_evidence_ids":[],"mechanisms":[]}]}`. Use empty arrays when
-annotations are absent and no extra keys.
-""".strip()
-
-
-def build_finding_synthesis_prompt(
-    payload: dict[str, Any],
-) -> tuple[str, str]:
-    result_set = (
-        payload.get("result_set")
-        if isinstance(payload.get("result_set"), dict)
-        else {}
-    )
-    candidate_rejection = (
-        payload.get("candidate_rejection")
-        if isinstance(payload.get("candidate_rejection"), dict)
-        else {}
-    )
-    rejection_reason = str(candidate_rejection.get("reason") or "").strip()
-    prompt_payload = {
-        "objective": (
-            payload.get("objective")
-            if isinstance(payload.get("objective"), dict)
-            else {}
-        ),
-        "result_set": {
-            key: result_set[key]
-            for key in (
-                "factors",
-                "outcome",
-                "primary_direction",
-                "total_evidence_count",
-                "is_condition_series",
-                "result_evidence",
-                "document_evidence_summaries",
-            )
-            if key in result_set
-        },
-        "paper_contributions": payload.get("paper_contributions") or [],
-        "context_evidence": payload.get("context_evidence") or [],
-    }
-    if rejection_reason:
-        prompt_payload["candidate_rejection"] = {"reason": rejection_reason}
-    input_json = json.dumps(
-        prompt_payload,
-        ensure_ascii=False,
-        separators=(",", ":"),
-    )
-    repair_contract = ""
-    if rejection_reason:
-        repair_contract = (
-            "Semantic repair required:\n"
-            f"- The previous candidate was rejected because: {rejection_reason}\n"
-            "- Return one corrected judgment only if it satisfies the original "
-            "Evidence contract; otherwise return an empty findings array.\n"
-            "- Return only ids present in `context_evidence`.\n\n"
-        )
-    user_prompt = (
-        "Judge assertion strength and optional context for this backend-owned "
-        "atomic result set.\n\n"
-        f"Input JSON:\n{input_json}\n\n"
-        f"{repair_contract}"
-        "Return only schema-valid JSON with a `findings` array. Return at most one "
-        "item containing assertion strength, context Evidence ids, and subordinate "
-        "mechanisms. Return `{\"findings\":[]}` when the supplied result does not "
-        "support a defensible Finding."
-    )
-    return _FINDING_SYNTHESIS_SYSTEM_PROMPT, user_prompt
