+++ backend/tests/unit/application/test_objective_analysis_workflow.py
@@ -418,7 +418,7 @@ def test_objective_analysis_uses_conservative_frame_batch_when_model_fails(
         collection_service=collection_service,
         objective_extractor=extractor,
     )
-    service.finding_synthesis_service.finding_extractor = extractor
+    service.finding_synthesis_service.assertion_judge = extractor
     service.source_artifact_repository.replace_collection_documents(
         collection_id,
         "build_test",
@@ -504,7 +504,7 @@ def test_objective_analysis_uses_deterministic_route_when_route_model_fails(
         collection_service=collection_service,
         objective_extractor=_ObjectiveExtractor(),
     )
-    service.finding_synthesis_service.finding_extractor = service._objective_extractor
+    service.finding_synthesis_service.assertion_judge = service._objective_extractor
     service.source_artifact_repository.replace_collection_documents(
         collection_id,
         "build_test",
@@ -577,7 +577,7 @@ def test_objective_analysis_uses_deterministic_route_when_route_model_fails(
     failing_extractor = _FailingRouteExtractor()
     service._objective_extractor = failing_extractor
     service._objective_evidence_router = failing_extractor
-    service.finding_synthesis_service.finding_extractor = failing_extractor
+    service.finding_synthesis_service.assertion_judge = failing_extractor
     artifacts = service.generate_objective_analysis_artifacts(
         collection_id, analysis
     )
@@ -601,7 +601,7 @@ def test_objective_analysis_does_not_mutate_active_objective_facts(
         collection_service=collection_service,
         objective_extractor=extractor,
     )
-    service.finding_synthesis_service.finding_extractor = extractor
+    service.finding_synthesis_service.assertion_judge = extractor
     service.source_artifact_repository.replace_collection_documents(
         collection_id,
         "build_test",
@@ -698,7 +698,7 @@ def test_queued_analysis_uses_its_source_build_objective_scope_after_rebuild(
         objective_extractor=extractor,
         objective_repository=repository,
     )
-    service.finding_synthesis_service.finding_extractor = extractor
+    service.finding_synthesis_service.assertion_judge = extractor
     service.source_artifact_repository.replace_collection_documents(
         collection_id,
         "build_test",
